Second weapon rework
Changes the attack, defense and itemdefense variables to be inherited from their respective superclass. This allows the three variables to be referenced by the Equipment class, which allows for item stat comparison.
Additionally, adds code to check for equipment with lower stats than the currently-equipped equipment of it's class. If lesser, the equipment will be "recycled" into gold. Addresses a part of #2