Turn Fixes
Fixes turns being executed by any button press.
This commit adds 'keypressed', which takes over the role of the old 'key'. 'key' has been repurposed into a String Array, which now holds all the possible control keys (w,a,s,d,space).
Act method in Player checks whether or not the user has pressed a key, and uses a for loop to cross-reference it with an array. If the key was a control key, the code executes as normal. If not, the code does nothing, preventing the turn from occurring from that keypress. Closes #4