Updates master to dev branch. (#19)
* Merge Conflict Aftermath 2 Elec Boog

Starts fixing the aftermath of all the failed merge conflicts. Removes some HEADS and DEVS and lines of = and < or >. Doesn't fix any issues.

* Starting to fix movement

Adds the framework for a new movement check, based on checking the state of 'keypressed' instead of doing independent checks.

* Fixes movement and fighting

Fixes the ability to move, interact with the world and fight monsters.