Updates master with latest dev stable. (#25)
* Shop Update pt1 - Music

Adds Music.class, a universal music controller that exists in every level. Music control is based on the player, and certain methods in the music class. Additionally, sound adjustments.

* Shop Update pt2 - Visual Progression

Starts to address #22, but does not complete it. Adds new sprites for turfs, and starting components for changing sprites based on difficulty. Currently does not compile, due to non-arbitrary integers being utilised.

* Shop Update pt3 - Fixing Doors

Fixes how door opening is handled. One, it's handled in movement method instead of act. Two, it fixes this new method of finding the door, it's difficulty level and then translating it over to the new door. Lastly, adds a new lose screen to DeathWorld.

* Shop Update pt4 - Fixing Visual Progression, and more.

This commit fixes the code that made doors inherit their difficulty from one another when opening/closing. Additionally, it adds a new class, "EndButton", which acts similarly to StartButton, but brings the player to the StartWorld instead. It is used in the end screens DeathWorld and EndWorld.